Now version is stored as double in Config

+++ src/main/java/me/i2000c/newalb/config/ReadWriteConfig.java
@@ -22,7 +22,7 @@ public class ReadWriteConfig extends ReadOnlyConfig{
     }
     
     @Override
-    public int getConfigVersion(){
+    public double getConfigVersion(){
         return INVALID_CONFIG_VERSION;
     }
 }
